backend/requirements.txt
# AgonX Backend Dependencies

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
aiomysql==0.2.0
alembic==1.13.2
pymysql==1.1.1

# Redis
redis==5.1.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# LangChain & LangGraph
langchain==0.3.1
langchain-core==0.3.6
langchain-community==0.3.0
langchain-openai==0.2.0
langgraph==0.2.28

# Vector Database
pymilvus==2.4.6

# Object Storage
minio==7.2.8

# Embedding Models
sentence-transformers==3.1.1

# Document Processing
pypdf==5.0.0
python-docx==1.1.2
unstructured==0.15.13

# PDF Export
fpdf2==2.7.9
reportlab==4.2.2

# HTTP Client
httpx==0.27.2
aiohttp==3.10.5

# Utils
python-dotenv==1.0.1
tenacity==9.0.0
tiktoken==0.7.0

# Reranker
FlagEmbedding==1.2.11
